Refactor calendar library under desktop package
